Move decorators and binding to new common bundle that the cdi.2.0 bundle will depend on
