+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/generated/
