+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/.settings/bndtools.core.prefs
@@ -1,0 +1,2 @@
+compileErrorAction=build
+eclipse.preferences.version=1
