+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,3 @@
+#Wed Aug 17 16:19:12 BST 2011
+eclipse.preferences.version=1
+encoding/<project>=UTF-8
