+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/bnd.bnd
@@ -14,8 +14,8 @@ bVersion=1.0
 javac.source: 1.8
 javac.target: 1.8
 
-Bundle-Name: com.ibm.ws.microprofile.metrics.cdi
-Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.cdi.2.0
+Bundle-Name: com.ibm.ws.microprofile.metrics.cdi.common
+Bundle-SymbolicName: com.ibm.ws.microprofile.metrics.cdi.common
 Bundle-Description: MicroProfile Metrics CDI Integration, version ${bVersion}
 
 Import-Package: \
@@ -23,25 +23,18 @@ Import-Package: \
 	*
 
 Export-Package: \
-	io.astefanutti.metrics.cdi,\
-	com.ibm.ws.microprofile.metrics.cdi.decorator,\
-	com.ibm.ws.microprofile.metrics.cdi.binding,\
-	com.ibm.ws.microprofile.metrics.cdi.interceptors,\
-	com.ibm.ws.microprofile.metrics.cdi.producer,\
-	com.ibm.ws.microprofile.metrics.cdi.helper
+    com.ibm.ws.microprofile.metrics.cdi.decorator,\
+    com.ibm.ws.microprofile.metrics.cdi.binding
 
 Private-Package: \
-    com.ibm.ws.microprofile.metrics.cdi.resources,\
-    com.ibm.ws.microprofile.metrics.cdi.helper
+    com.ibm.ws.microprofile.metrics.cdi.resources
 
 Include-Resource: \
     META-INF=resources/META-INF
 
 WS-TraceGroup: METRICS
+publish.wlp.jar.disabled: true
 
--dsannotations-inherit: true
--dsannotations: \
-	io.astefanutti.metrics.cdi.MetricsExtension
 
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 
@@ -56,14 +49,10 @@ Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 	com.ibm.websphere.org.eclipse.microprofile.config.1.3;version=latest,\
 	com.ibm.ws.classloading;version=latest,\
 	com.ibm.ws.kernel.service;version=latest,\
-	com.ibm.ws.microprofile.metrics.2.0;version=latest,\
-	com.ibm.websphere.org.eclipse.microprofile.metrics.2.0;version=latest,\
 	com.ibm.websphere.org.osgi.core;version=latest,\
 	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
 	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
     com.ibm.websphere.javaee.interceptor.1.2;version=latest,\
     com.ibm.websphere.javaee.annotation.1.2;version=latest,\
-    com.ibm.websphere.javaee.concurrent.1.0;version=latest,\
-    com.ibm.ws.microprofile.metrics.cdi.common,\
-    com.ibm.ws.microprofile.metrics.common;version=latest
+    com.ibm.websphere.javaee.concurrent.1.0;version=latest
