+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/src/com/ibm/ws/microprofile/metrics/cdi/binding/MetricsBinding.java
@@ -21,7 +21,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.ibm.ws.microprofile.metrics.cdi.interceptors;
+package com.ibm.ws.microprofile.metrics.cdi.binding;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
