+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/src/com/ibm/ws/microprofile/metrics/cdi/decorator/AnnotatedTypeDecorator.java
@@ -21,7 +21,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.ibm.ws.microprofile.metrics.cdi;
+package com.ibm.ws.microprofile.metrics.cdi.decorator;
 
 import java.lang.annotation.Annotation;
 import java.util.Collections;
