+++ dev/com.ibm.ws.microprofile.metrics.cdi.common/src/com/ibm/ws/microprofile/metrics/cdi/decorator/package-info.java
@@ -1,0 +1,15 @@
+/*******************************************************************************
+ * Copyright (c) 2019 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+/**
+ *
+ */
+@org.osgi.annotation.versioning.Version("1.0")
+package com.ibm.ws.microprofile.metrics.cdi.decorator;
